+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Refactor function expression to use `FunctionRender` interface
 - Complete `POSITION`, add private method encapsulation for all functions, and rewrite `COALESCE`, `NULLIF`, `CAST`
+- Implement `CASE` expression function
 
 ## 1.1.0